XIVY-12549 Fix inserted string for functionbrowser
